Remove legacy executor/plan code
