+++ src/conductor/execution/executor.py
@@ -3,14 +3,15 @@ import time
 import collections
 import os
 import signal
-from typing import Dict, List, Iterable, Tuple
+from typing import Dict, List, Iterable, Tuple, Deque
 
 from conductor.context import Context
 from conductor.errors import ConductorError, ConductorAbort
+from conductor.execution.ops.operation import Operation
 from conductor.execution.plan import ExecutionPlan
-from conductor.execution.task import ExecutingTask
 from conductor.execution.task_state import TaskState
 from conductor.task_types.base import TaskExecutionHandle
+from conductor.task_identifier import TaskIdentifier
 from conductor.utils.colored_output import (
     print_bold,
     print_cyan,
@@ -23,58 +24,58 @@ from conductor.utils.time import time_to_readable_string
 
 
 class _ReadyToRunQueue:
-    def __init__(self):
-        self._sequential_tasks = collections.deque()
-        self._parallel_tasks = collections.deque()
+    def __init__(self) -> None:
+        self._sequential_ops: Deque[Operation] = collections.deque()
+        self._parallel_ops: Deque[Operation] = collections.deque()
 
-    def load(self, initial_tasks: Iterable[ExecutingTask]):
-        for task in initial_tasks:
-            self.enqueue_task(task)
+    def load(self, initial_ops: Iterable[Operation]) -> None:
+        for op in initial_ops:
+            self.enqueue_op(op)
 
-    def has_tasks(self) -> bool:
-        return (len(self._sequential_tasks) > 0) or (len(self._parallel_tasks) > 0)
+    def has_ops(self) -> bool:
+        return (len(self._sequential_ops) > 0) or (len(self._parallel_ops) > 0)
 
-    def has_parallelizable_tasks(self) -> bool:
-        return len(self._parallel_tasks) > 0
+    def has_parallelizable_ops(self) -> bool:
+        return len(self._parallel_ops) > 0
 
-    def enqueue_task(self, task: ExecutingTask) -> None:
-        if task.task.parallelizable:
-            self._parallel_tasks.append(task)
+    def enqueue_op(self, op: Operation) -> None:
+        if op.parallelizable:
+            self._parallel_ops.append(op)
         else:
-            self._sequential_tasks.append(task)
+            self._sequential_ops.append(op)
 
-    def dequeue_next(self) -> ExecutingTask:
-        if self.has_parallelizable_tasks():
-            return self._parallel_tasks.popleft()
+    def dequeue_next(self) -> Operation:
+        if self.has_parallelizable_ops():
+            return self._parallel_ops.popleft()
         else:
-            return self._sequential_tasks.popleft()
+            return self._sequential_ops.popleft()
 
     def clear(self) -> None:
-        self._sequential_tasks.clear()
-        self._parallel_tasks.clear()
+        self._sequential_ops.clear()
+        self._parallel_ops.clear()
 
 
-class _InflightTasks:
-    def __init__(self):
-        self._processes: Dict[int, Tuple[TaskExecutionHandle, ExecutingTask]] = {}
-        self._sync_tasks: List[Tuple[TaskExecutionHandle, ExecutingTask]] = []
+class _InflightOperations:
+    def __init__(self) -> None:
+        self._processes: Dict[int, Tuple[TaskExecutionHandle, Operation]] = {}
+        self._sync_ops: List[Tuple[TaskExecutionHandle, Operation]] = []
 
-    def add_task(self, handle: TaskExecutionHandle, task: ExecutingTask) -> None:
+    def add_op(self, handle: TaskExecutionHandle, op: Operation) -> None:
         if handle.is_sync:
-            self._sync_tasks.append((handle, task))
+            self._sync_ops.append((handle, op))
         else:
             assert handle.pid is not None
-            self._processes[handle.pid] = (handle, task)
+            self._processes[handle.pid] = (handle, op)
 
     def __len__(self) -> int:
-        return len(self._processes) + len(self._sync_tasks)
+        return len(self._processes) + len(self._sync_ops)
 
-    def has_sync_tasks(self) -> bool:
-        return len(self._sync_tasks) > 0
+    def has_sync_ops(self) -> bool:
+        return len(self._sync_ops) > 0
 
-    def wait_for_next_task(self) -> Tuple[TaskExecutionHandle, ExecutingTask]:
-        if len(self._sync_tasks) > 0:
-            return self._sync_tasks.pop()
+    def wait_for_next_op(self) -> Tuple[TaskExecutionHandle, Operation]:
+        if len(self._sync_ops) > 0:
+            return self._sync_ops.pop()
 
         # Wait for the next child process to finish.
         while True:
@@ -107,18 +108,18 @@ class _InflightTasks:
     def clear(self) -> None:
         self.terminate_processes()
         self._processes.clear()
-        self._sync_tasks.clear()
+        self._sync_ops.clear()
 
 
 class Executor:
-    def __init__(self, execution_slots: int):
+    def __init__(self, execution_slots: int) -> None:
         assert execution_slots > 0
         self._slots = execution_slots
         self._available_slots = list(reversed(range(self._slots)))
 
         self._ready_to_run = _ReadyToRunQueue()
-        self._inflight_tasks = _InflightTasks()
-        self._completed_tasks: List[ExecutingTask] = []
+        self._inflight_ops = _InflightOperations()
+        self._completed_ops: List[Operation] = []
         self._running_parallel = False
         self._num_tasks_to_run = 0
         self._num_tasks_dequeued = 0
@@ -128,32 +129,31 @@ class Executor:
     ):
         try:
             self._reset()
+            plan.reset_waiting_on()
             self._num_tasks_to_run = plan.num_tasks_to_run
             start = time.time()
 
             # 1. Print out any cached tasks.
             for cached_task in plan.cached_tasks:
                 print_cyan(
-                    "✓ Using cached results for {}.".format(
-                        str(cached_task.task.identifier)
-                    )
+                    "✓ Using cached results for {}.".format(str(cached_task.identifier))
                 )
 
-            # 2. Run the tasks as they become eligible for execution.
+            # 2. Run the operations as they become eligible for execution.
             should_stop = False
-            self._ready_to_run.load(plan.initial_tasks)
+            self._ready_to_run.load(plan.initial_ops)
             with SigchldHelper.instance().track():
-                while self._ready_to_run.has_tasks() or len(self._inflight_tasks) > 0:
-                    should_stop = self._launch_tasks_if_able(ctx, stop_on_first_error)
+                while self._ready_to_run.has_ops() or len(self._inflight_ops) > 0:
+                    should_stop = self._launch_ops_if_able(ctx, stop_on_first_error)
                     if should_stop:
                         break
 
-                    if len(self._inflight_tasks) == 0:
-                        # There may be no in-flight tasks if the last ready-to-run
-                        # task failed or was skipped.
+                    if len(self._inflight_ops) == 0:
+                        # There may be no in-flight ops if the last ready-to-run
+                        # op failed or was skipped.
                         continue
 
-                    should_stop = self._wait_for_next_inflight_task(
+                    should_stop = self._wait_for_next_inflight_op(
                         ctx, stop_on_first_error
                     )
                     if should_stop:
@@ -161,13 +161,13 @@ class Executor:
 
             # Only has an effect if we exited the loop above early due to
             # encountering an error.
-            self._inflight_tasks.terminate_processes()
+            self._inflight_ops.terminate_processes()
 
             # 3. Report the results.
             self._report_execution_results(plan, elapsed_time=(time.time() - start))
 
         except ConductorAbort:
-            self._inflight_tasks.terminate_processes()
+            self._inflight_ops.terminate_processes()
             elapsed = time.time() - start
             print()
             print_yellow(
@@ -179,38 +179,39 @@ class Executor:
 
     def _reset(self) -> None:
         self._ready_to_run.clear()
-        self._completed_tasks.clear()
-        self._inflight_tasks.clear()
+        self._completed_ops.clear()
+        self._inflight_ops.clear()
         self._running_parallel = False
         self._available_slots = list(reversed(range(self._slots)))
         self._num_tasks_to_run = 0
         self._num_tasks_dequeued = 0
 
-    def _launch_tasks_if_able(self, ctx: Context, stop_on_first_error: bool) -> bool:
+    def _launch_ops_if_able(self, ctx: Context, stop_on_first_error: bool) -> bool:
         """
-        Launches as many tasks as possible while respecting the task's
+        Launches as many operations as possible while respecting the operation's
         parallelization setting and the number of execution slots available.
 
         Returns `True` if an error occurs and `stop_on_first_error` is set to `True.
         """
         while True:
-            can_launch_any_ready_task = (
-                self._ready_to_run.has_tasks() and len(self._inflight_tasks) == 0
+            can_launch_any_ready_op = (
+                self._ready_to_run.has_ops() and len(self._inflight_ops) == 0
             )
-            can_launch_parallelizable_task = (
+            can_launch_parallelizable_op = (
                 self._running_parallel
-                and len(self._inflight_tasks) < self._slots
-                and self._ready_to_run.has_parallelizable_tasks()
+                and len(self._inflight_ops) < self._slots
+                and self._ready_to_run.has_parallelizable_ops()
             )
 
-            if not can_launch_any_ready_task and not can_launch_parallelizable_task:
+            if not can_launch_any_ready_op and not can_launch_parallelizable_op:
                 break
 
             # Parallelizable tasks are prioritized (dequeued first).
-            next_task = self._ready_to_run.dequeue_next()
+            next_op = self._ready_to_run.dequeue_next()
             prev_running_parallel = self._running_parallel
-            self._running_parallel = next_task.task.parallelizable
-            self._num_tasks_dequeued += 1
+            self._running_parallel = next_op.parallelizable
+            if next_op.main_task is not None:
+                self._num_tasks_dequeued += 1
 
             # For output cosmetics. We add empty lines between task outputs to
             # help distinguish their outputs. But this extra empty line is not
@@ -221,107 +222,117 @@ class Executor:
                 prev_running_parallel and self._running_parallel and self._slots > 1
             )
 
-            if not next_task.exe_deps_succeeded():
+            if not next_op.exe_deps_succeeded():
                 # At least one dependency failed, so we need to skip this task.
-                if not avoid_leading_newline:
-                    print()
-                print_yellow(
-                    "✱ Skipping {}. {}".format(
-                        str(next_task.task.identifier), self._get_progress_string()
+                if next_op.main_task is not None:
+                    # Print information about the task that is being skipped.
+                    if not avoid_leading_newline:
+                        print()
+                    print_yellow(
+                        "✱ Skipping {}. {}".format(
+                            str(next_op.main_task.identifier),
+                            self._get_progress_string(),
+                        )
                     )
-                )
-                next_task.set_state(TaskState.SKIPPED)
-                self._process_finished_task(next_task)
+                next_op.set_state(TaskState.SKIPPED)
+                self._process_finished_op(next_op)
             else:
-                if not avoid_leading_newline:
-                    print()
-                print_cyan(
-                    "✱ Running {}... {}".format(
-                        str(next_task.task.identifier), self._get_progress_string()
+                if next_op.main_task is not None:
+                    if not avoid_leading_newline:
+                        print()
+                    print_cyan(
+                        "✱ Running {}... {}".format(
+                            str(next_op.main_task.identifier),
+                            self._get_progress_string(),
+                        )
                     )
-                )
                 try:
                     slot = (
                         self._available_slots[-1]
                         if self._running_parallel and self._slots > 1
                         else None
                     )
-                    handle = next_task.task.start_execution(ctx, slot)
+                    handle = next_op.start_execution(ctx, slot)
                     handle.slot = slot
-                    self._inflight_tasks.add_task(handle, next_task)
+                    self._inflight_ops.add_op(handle, next_op)
                     if slot is not None:
                         self._available_slots.pop()
                 except ConductorAbort:
-                    next_task.set_state(TaskState.ABORTED)
+                    next_op.set_state(TaskState.ABORTED)
                     # N.B. A slot may be leaked here, but it does not matter
                     # because we are aborting the execution.
                     raise
                 except ConductorError as ex:
-                    next_task.store_error(ex)
-                    next_task.set_state(TaskState.FAILED)
-                    self._process_finished_task(next_task)
-                    self._print_task_failed(next_task)
+                    next_op.store_error(ex)
+                    next_op.set_state(TaskState.FAILED)
+                    self._process_finished_op(next_op)
+                    self._print_op_failed(next_op)
                     if stop_on_first_error:
                         return True
 
         return False
 
-    def _wait_for_next_inflight_task(
+    def _wait_for_next_inflight_op(
         self, ctx: Context, stop_on_first_error: bool
     ) -> bool:
         """
-        Waits for the next inflight task to complete (and processes it).
+        Waits for the next inflight operation to complete (and processes it).
 
         Returns `True` if an error occurs and `stop_on_first_error` is set to `True.
         """
-        assert len(self._inflight_tasks) > 0
+        assert len(self._inflight_ops) > 0
 
         error_occurred = False
-        handle, task = self._inflight_tasks.wait_for_next_task()
+        handle, op = self._inflight_ops.wait_for_next_op()
         try:
-            task.task.finish_execution(handle, ctx)
-            task.set_state(TaskState.SUCCEEDED)
-            print_green("✓ {} completed successfully.".format(task.task.identifier))
+            op.finish_execution(handle, ctx)
+            op.set_state(TaskState.SUCCEEDED)
+            if op.main_task is not None:
+                print_green(
+                    "✓ {} completed successfully.".format(op.main_task.identifier)
+                )
         except ConductorAbort:
-            task.set_state(TaskState.ABORTED)
+            op.set_state(TaskState.ABORTED)
             # N.B. A slot may be leaked here, but it does not matter because we
             # are aborting the execution.
             raise
         except ConductorError as ex:
-            task.store_error(ex)
-            task.set_state(TaskState.FAILED)
+            op.store_error(ex)
+            op.set_state(TaskState.FAILED)
             error_occurred = True
-            self._print_task_failed(task)
+            self._print_op_failed(op)
 
         if handle.slot is not None:
             self._available_slots.append(handle.slot)
-        self._process_finished_task(task)
+        self._process_finished_op(op)
 
         return error_occurred and stop_on_first_error
 
-    def _process_finished_task(self, finished_task: ExecutingTask) -> None:
-        self._completed_tasks.append(finished_task)
-        finished_task.decrement_deps_of_waiting_on()
-        for dep_of in finished_task.deps_of:
+    def _process_finished_op(self, finished_op: Operation) -> None:
+        self._completed_ops.append(finished_op)
+        finished_op.decrement_deps_of_waiting_on()
+        for dep_of in finished_op.deps_of:
             if dep_of.waiting_on > 0:
                 continue
-            self._ready_to_run.enqueue_task(dep_of)
+            self._ready_to_run.enqueue_op(dep_of)
 
     def _report_execution_results(self, plan: ExecutionPlan, elapsed_time: float):
-        all_succeeded = all(map(lambda task: task.succeeded(), self._completed_tasks))
-        # We executed at least one task.
-        # The main task we wanted to run should always be the last
-        # completed task (its dependencies must be executed first).
-        main_task_executed = (
-            len(self._completed_tasks) > 0
-            and self._completed_tasks[-1].task.identifier
-            == plan.task_to_run.task.identifier
+        all_succeeded = all(map(lambda op: op.succeeded(), self._completed_ops))
+        main_task_executed = any(
+            [
+                op.main_task is not None
+                and op.main_task.identifier == plan.task_to_run.identifier
+                for op in self._completed_ops
+            ]
         )
-        # We did not run any tasks, so the task we wanted to run
-        # must have been cached.
-        main_task_cached = (
-            len(self._completed_tasks) == 0
-            and plan.task_to_run.state == TaskState.SUCCEEDED_CACHED
+
+        # We did not run any ops, so the task we wanted to run must have been
+        # cached.
+        main_task_cached = len(self._completed_ops) == 0 and any(
+            [
+                task.identifier == plan.task_to_run.identifier
+                for task in plan.cached_tasks
+            ]
         )
 
         # Print the final execution result (succeeded or failed).
@@ -333,14 +344,16 @@ class Executor:
 
         else:
             # At least one task must have failed.
-            failed_tasks: List[ExecutingTask] = []
-            skipped_tasks: List[ExecutingTask] = []
-            for exe_task in self._completed_tasks:
-                if exe_task.state == TaskState.SKIPPED:
-                    skipped_tasks.append(exe_task)
-                elif exe_task.state == TaskState.FAILED:
-                    failed_tasks.append(exe_task)
-            assert len(failed_tasks) > 0
+            failed_task_ops: List[Operation] = []
+            skipped_tasks: List[TaskIdentifier] = []
+            for op in self._completed_ops:
+                if op.main_task is None:
+                    continue
+                if op.state == TaskState.SKIPPED:
+                    skipped_tasks.append(op.main_task.identifier)
+                elif op.state == TaskState.FAILED:
+                    failed_task_ops.append(op)
+            assert len(failed_task_ops) > 0
             print()
             print_red(
                 "🔴 Task failed. {}".format(
@@ -350,8 +363,9 @@ class Executor:
             )
             print()
             print_bold("Failed task(s):")
-            for failed in failed_tasks:
-                print("  {}".format(failed.task.identifier))
+            for failed in failed_task_ops:
+                assert failed.main_task is not None
+                print("  {}".format(failed.main_task.identifier))
                 assert failed.stored_error is not None
                 print(
                     "    {}".format(
@@ -362,14 +376,16 @@ class Executor:
             if len(skipped_tasks) > 0:
                 print_bold("Skipped task(s) (one or more dependencies failed):")
                 for skipped in skipped_tasks:
-                    print("  {}".format(skipped.task.identifier))
+                    print("  {}".format(skipped))
                 print()
 
-            assert failed_tasks[0].stored_error is not None
-            raise failed_tasks[0].stored_error
+            assert failed_task_ops[0].stored_error is not None
+            raise failed_task_ops[0].stored_error
 
-    def _print_task_failed(self, task: ExecutingTask):
-        print_red("✘ {} failed.".format(task.task.identifier))
+    def _print_op_failed(self, op: Operation):
+        if op.main_task is None:
+            return
+        print_red("✘ {} failed.".format(op.main_task.identifier))
 
     def _get_progress_string(self):
         return "({}/{})".format(str(self._num_tasks_dequeued), self._num_tasks_to_run)
